coverity#1242737 Result is not floating-point

+++ svx/source/unodraw/UnoGraphicExporter.cxx
@@ -687,8 +687,8 @@ bool GraphicExporter::GetGraphic( ExportSettings& rSettings, Graphic& aGraphic, 
                         else
                             nHeightPix = aSizePix.Height();
 
-                        double fWidthDif = aSizePix.Width() / nWidthPix;
-                        double fHeightDif = aSizePix.Height() / nHeightPix;
+                        double fWidthDif = (double)aSizePix.Width() / nWidthPix;
+                        double fHeightDif = (double)aSizePix.Height() / nHeightPix;
 
                         if (fWidthDif > fHeightDif)
                             nHeightPix = static_cast<long>(aSizePix.Height() / fWidthDif);
